==> 00-setup-Kang18.qmd <==
# Mereu18

## Preamble

```{r}
#| label: load-libs
library(scran)
library(scater)
library(patchwork)
library(ExperimentHub)
```

## Load data

```{r}
#| label: load-data
# load SCE from EH
eh <- ExperimentHub()
q <- query(eh, "Kang18_8vs8")
sce <- eh[[q$ah_id]]
```

## Wrangling

```{r}
# keep controls only
sce <- sce[, sce$stim == "ctrl"]
# drop unassigned cells & multiplets
sce <- sce[, !is.na(sce$cell)]
sce <- sce[, sce$multiplets == "singlet"]
# sparsify counts
counts(sce) <- as(counts(sce), "dgCMatrix")
# standardize metadata
colData(sce) <- DataFrame(
  cluster = sce$cell,
  batch = factor(sce$ind))
# tabulate cell counts
table(sce$cluster, sce$batch)
```

```{r}
#| label: calc-pca
sce <- logNormCounts(sce)
tbl <- modelGeneVar(sce, block = sce$batch)
hvg <- rowData(sce)$hvg <- tbl$bio > 0
sce <- runPCA(sce, subset_row = hvg)
```

```{r}
#| label: plot-thm
thm <- list(
  theme(aspect.ratio = 1, legend.key.size = unit(0.5, "lines")),
  guides(col = guide_legend(override.aes = list(alpha = 1, size = 2))))
```

```{r fig.width = 10, fig.height = 4}
#| label: plot-pca
plotReducedDim(sce, "PCA", colour_by = "batch", point_size = 0.1) +
plotReducedDim(sce, "PCA", colour_by = "cluster", point_size = 0.1) +
plot_layout() & thm
```

## Appendix

```{r}
#| label: save-data
saveRDS(sce, "outs/Kang18.rds")
```
