==> 00-setup-Mereu20.qmd <==
# Mereu20

## Preamble

```{r}
#| label: load-libs
library(scran)
library(scater)
library(patchwork)
library(SingleCellExperiment)
```

## Load data

```{r}
#| label: load-data
# download dataset
con <- url(paste0(
  "https://www.dropbox.com/s/i8mwmyymchx8mn8/",
  "sce.all_classified.technologies.RData?raw=1"))
options(timeout = 180)
sce <- get(load(con))
close(con)
sce
```

## Wrangling

```{r}
# standardize metadata
i <- grep("nnet2", names(colData(sce)))
names(colData(sce))[i] <- "cluster"
sce$cluster <- droplevels(sce$cluster)
# tabulate cell counts
table(sce$batch, sce$cluster)
```

The published data has been integrated across batches using `harmony`.
We do another round of PCA w/o correction for later comparison comparison.

```{r}
# basic feature selection
tbl <- modelGeneVar(sce, block = sce$batch)
hvg <- rowData(sce)$hvg <- tbl$bio > 0
# compute uncorrected PCA
reducedDimNames(sce)[2] <- "HARMONY"
dim <- colnames(reducedDim(sce, 2))
dim <- gsub("_", "", dim)
colnames(reducedDim(sce, 2)) <- dim
sce <- runPCA(sce, subset_row = hvg)
```

```{r}
#| label: plot-thm
thm <- list(
  theme(aspect.ratio = 1, legend.key.size = unit(0.5, "lines")),
  guides(col = guide_legend(override.aes = list(alpha = 1, size = 2))))
```

```{r fig.width = 9, fig.height = 4}
#| label: plot-dr1
plotReducedDim(sce, "HARMONY", colour_by = "cluster", point_size = 0.1) +
plotReducedDim(sce, "UMAP", colour_by = "cluster", point_size = 0.1) +
plot_layout(guides = "collect") & thm
```

```{r fig.width = 10, fig.height = 4}
#| label: plot-dr2
plotReducedDim(sce, "PCA", colour_by = "cluster", point_size = 0.1) +
plotReducedDim(sce, "PCA", colour_by = "batch", point_size = 0.1) +
plot_layout() & thm
```

## Appendix

```{r}
#| label: save-data
saveRDS(sce, "outs/Mereu20.rds")
```
